// packages/portal/backend/src/Types.ts
export enum PersonKind {
    NONE = "",
    STUDENT = "student",
    WITHDRAWN = "withdrawn", // typically a student who has left the class
    ADMINSTAFF = "adminstaff",
    ADMIN = "admin",
    STAFF = "staff"
}

export interface Auth {
    personId: string; // invariant
    token: string | null;
}

// NOTE: Intentionally not linked to Repository (see docs at top of file)
export interface Deliverable {
    readonly id: string; // primary key; invariant. this is the shortname of the deliverable (e.g., d1)
    URL: string; // links to the public deliverable description

    openTimestamp: number;
    closeTimestamp: number;
    gradesReleased: boolean; // whether students can see their grades

    visibleToStudents: boolean; // whether students even see the column

    rubric: any; // captures rubric-specific definitions
    // custom: any; // {}; not used by the default implementation, but useful for extension (e.g., schemas)
    custom: {
        rubric?: any, // CS340REST
        assignment?: AssignmentInfo // AssignmentController
        // courseWeight?: any, // AssignmentController // TODO: make into assignment.courseWeight
        // seedRepoURL?: any, // RubricController // TODO: make into rubric.seedRepoURL
        // seedRepoPath?: any, // RubricController // TODO: make into rubric.seedRepoPath
        // mainFilePath?: any // AssignmentController // TODO: make into assignment.mainFilePath
    };

    lateAutoTest: boolean; // whether the deliv can be executed after the deadline
    shouldAutoTest: boolean; // whether the deliv will use AutoTest
    autotest: AutoTestConfig;

    // these options are only set if shouldProvision is true
    shouldProvision: boolean; // whether the deliv is for provisioning at all; if not, the fields below are not needed
    repoPrefix: string | null; // prefix for repo names (e.g., project_ or d1_)
    teamPrefix: string | null; // prefix for team names (e.g., pTeam_ or d1Team_)
    importURL: string | null; // URL that should be cloned for the repos to be provisioned
    teamMinSize: number;
    teamMaxSize: number;
    teamSameLab: boolean;
    teamStudentsForm: boolean;
}
// ...
export interface Repository {
    /**
     * The name of the repo; must be unique locally and on GitHub.
     */
    readonly id: string; // invariant
    /**
     * The deliverable the repository was provisioned for. This does not modify AutoTest
     * but is used to track provisioning.
     */
    readonly delivId: string; // invariant

    /**
     * URL for project in version control system; null if not yet created.
     *
     * TODO: remove; make githubProvisioned non-optional and use that instead.
     */
    URL: string | null;

    /**
     * git clone URL for project; null if not yet created
     *
     * TODO: make sure this is not used to check if a repo has been provisioned.
     */
    cloneURL: string | null;
    teamIds: string[]; // Team.id[] - foreign key

    /**
     * Has the repo been provisioned on GitHub.
     *
     * TODO: make this non-optional.
     */
    githubProvisioned?: boolean;

    /**
     * Is the repo student-writeable on GitHub.
     *
     * TODO: make this non-optional.
     */
    githubWriteable?: boolean;

    custom: { // rather than having custom be .any, this allows courses to make sure they do not clash on their .custom parameters
        githubCreated?: boolean, // TODO: remove
        githubReleased?: boolean, // TODO: remove

        // SDMM-specific; leave them until the SDMM is cancelled
        d0enabled?: boolean, // SDMM
        d1enabled?: boolean, // SDMM
        d2enabled?: boolean, // SDMM
        d3enabled?: boolean  // SDMM
        sddmD3pr?: boolean, // SDMM
    };
}

/**
 * This is just a placeholder type to hold course-level data that can change.
 * (in contrast to course-level static data in the .env file)
 */
export interface Course {
    readonly id: string; // invariant; this is the name of the course
    defaultDeliverableId: string | null; // Deliverable.id foreign key
    custom: {
        status?: string
    };
}

export enum AuditLabel {
    COURSE = "Course",
    DELIVERABLE = "Deliverable",
    REPOSITORY = "Repository",
    TEAM = "TEAM",
    TEAM_ADMIN = "TeamAdmin", // Created / updated by admin
    TEAM_STUDENT = "TeamStudent", // Created / updated by student
    GRADE_ADMIN = "GradeAdmin", // Created / updated by admin
    GRADE_CHANGE = "Grade_Change",
    GRADE_AUTOTEST = "GradeAutotest",
    REPO_PROVISION = "RepositoryProvision",
    REPO_RELEASE = "RepositoryRelease",
    CLASSLIST_UPLOAD = "Classlist_Upload",
    CLASSLIST_PRUNE = "Classlist_Prune"
}

export interface AuditEvent {
    label: string;
    timestamp: number;
    personId: string;
    before: object | null;
    after: object | null;
    custom: object; // enables easier querying
}

export interface Grade {
    // this should be the personId associated with the repo, not a staff who invoked it!
    readonly personId: string; // Person.id; grades are really on repos, but we only care about them by person
    readonly delivId: string; // Deliverable.id - foreign key // could be a Deliverable, but this is just easier

    score: number;
    comment: string;
    timestamp: number;

    urlName: string | null; // name associated with URL (e.g., project name)
    URL: string | null; // link to commit, if appropriate or repoUrl if not

    custom: any; // {}; not used by the default implementation, but useful for extension (e.g., custom grade values)
    /*
    custom: { // rather than having custom be .any, this allows courses to make sure they do not clash on their .custom parameters
        sdmmStatus?: boolean

        // questions?: any, // AssignmentController // TODO: make into assignment.questions
        // assignmentID?: any, // AssignmentController // TODO: make into assignment.id
        // studentID?: any, // AssignmentController // TODO: make into assignment.personId
        // released?: any, // AssignmentController // TODO: make into assignment.released
        assignmentGrade?: AssignmentGrade
    };
     */
}